requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.20.0
fpdf2>=2.7.8
numpy>=1.25.0
textwrap3>=0.9.2
